fix(ServerHealthContext): prevents reload loop on 500 errors.
The window.location.reload function that was added to make sure the page
recovered completely from server failure was causing infinite loading
loops when a 500 error was encountered. Moving this reload to only
occur when the modal dialog is closed, prevents the loop from occurring
and makes sure that the page is in a good state when the server
recovers.