feat(cli): Adds auto login end point for cli usage
The cli code enables a cli-mode which exposes an /api/auth/cli-login end
point. This endpoint accepts a token which has been pre-generated by the
cli code. The token is associated with a session which has been placed
in the session table. The session is determined by grabbing the user
from the server process.