fix: Generates random secret key for cli
When first installed, the cli code will not have a secret key in the
config file. The server launch generates a secret key, but this is after
the point where the cli needs it, so the cli is now generating the
secret key that will be used in both cases.